refactor: isolate api from utils
