+++ src/components/chat/ChannelInfoModal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ModalBase from '../common/ModalBase';
 import { Typography } from '@mui/material';
-import { TChannelProps } from '../../utils/chat/type';
+import { TChannelProps } from '../../api/chat/type';
 import SubsButton from '../search/SubsButton';
 import { useNavigate } from 'react-router-dom';
 type TChannelInfoModalProps = {
