+++ src/components/chat/MessagesPane.tsx
@@ -6,10 +6,10 @@ import MessageBubble from './MessageBubble';
 import MessageInput from './MessageInput';
 import AiPannel from '../ai/AiPannel';
 import { useChatStore, useSocket } from '../../state/chat';
-import { TMessageProps } from '../../utils/chat/type';
+import { TMessageProps } from '../../api/chat/type';
 import { useParams } from 'react-router-dom';
 import { useInfiniteQuery } from 'react-query';
-import { chatMessageQuery } from '../../utils/chat/query';
+import { chatMessageQuery } from '../../api/chat/query';
 import { useUserStore } from '../../state/user';
 import UserAvatar from '../user/UserAvatar';
 import { parseRawDateAndTime } from '../../utils/common/function';
