+++ src/components/user/UserAvatar.tsx
@@ -1,5 +1,5 @@
 import { Box } from '@mui/joy';
-import { TUserProps } from '../../utils/user/type';
+import { TUserProps } from '../../api/user/type';
 
 interface TUserAvatarProps {
   user: TUserProps | null;
