+++ src/pages/settings/SettingsPage.tsx
@@ -10,7 +10,7 @@ import {
 import { useUserStore } from '../../state/user';
 import EditIcon from '@mui/icons-material/Edit';
 import { useState } from 'react';
-import userApi from '../../utils/user/api';
+import userApi from '../../api/user/api';
 import Divider from '@mui/joy/Divider';
 const Settings = () => {
   const { nickname, profileUrl, email, setNickname } = useUserStore();
